Improved plugin inject —poll-files
With —poll-files, there could be a race condition. When a file is copied and tsp -P inject tries to load it at the same time, there is a race condition. The file can be locked for write by the copy operation and tsp cannot read it. Now, in case of error while loading a file, the plugin inject retries three times, at polling interval.