DektecInputPlugin:  Added support for demodulation frequency
Added --frequency, --satellite-frequency, and --lnb options to Dektec input plugin, which are effectively copies of similar functionality that already exists in the Dektec output plugin, along with making changes to the code to make use of these options.  Prior to making these changes, the Dektec input plugin didn't work with Dektec satellite receivers, such as the DTA-2137C.

Note:  Currently, the UHF and VHF options are not enabled, but they could be enabled, as they are with the Dektec output plugin, if someone has an appropriate Dektec module with which to test.  In addition, some of the code for handling frequencies was borrowed from the Dektec output plugin.

Comments:
-- src/libtsduck/tsDektecInputPlugin.cpp:
a) Added --frequency, --satellite-frequency, and --lnb command-line options.
b) Added code to parse new options in getOptions().
c) In start() method, made the following changes:
  1.  Prior to putting the channel in receive mode, now putting it into idle mode first and also clearing FIFO and flags--this mirrors how things are done in the DtRecord Dektec sample.
  2.  If the demodulation frequency is non-zero, it sets the tuner frequency and demodulation settings prior to putting the channel into receive mode.  Currently, it uses some defaults for the demodulation settings (QPSK, etc).  The Dektec output plugin provides the ability to use multiple modulation settings in an similar fashion, and equivalent support could be added to the input plugin in a future commit.

Signed-off-by: Aaron Levinson <alevinsn_dev@levland.net>