Added hasSplicingPoint() and getSplicingPoint() methods to TSPacket class, which expose the Splicing Point flag and the splice countdown, if specified, in a TS packet's adaptation field
Note:  This was added with the hope that I would be able to use the information in the rmsplice plug-in, but the TS I was working with didn't have the Splicing Point flag set in any packets.  In addition, the various offset methods now have a lot of code in common.  The various offset methods would benefit from code consolidation by moving them into a single method that is called once and stores the various offsets in member variables in the TSPacket class.