Fixes ugly optimization
When packet processing starts, in this moment the value of _lateIndex is 0. And when the first packet is queued, this value changes to 1. However, some optmization breaks this, and only after the flush of the queue the value returns to 1.

So one simple solution to fix this trouble is to set it to 1 at starting time. In any case, when the loop empties the queue, the value is set to 1 (see last line of function fillPacket() ).

This resolves the ugly output of initial packets.