Reporting: Add support for using OutputDebugString() on Windows if TS_DEBUG_LOG defined
If TS_DEBUG_LOG is defined when building on Windows,
ts::AsyncReport::writeLog() will also output log messages to the
debugger console using OutputDebugStringA().  This can be useful when
debugging, because log messages typically are output to the console in
an asynchronous fashion, and in addition, if there are many log
messages delivered within a short period of time, sometimes it can
overload the log queueing mechanism, resulting in dropped log
messages.  This define should not be enabled for production builds, as
the use of OutputDebugStringA() can impact performance.  Additionally,
if the user sets the TS_DEBUG_LOG environment variable to 1, it will
cause the TS_DEBUG_LOG define to be turned on for all tsduck projects.