ApplicationSharedLibrary: Add prefix support in fallback code when loading library from filename
While the filename prefix (as in "tsplugin_") is supported in the earlier code in the constructor, it wasn't being used in the fallback code that attempts to load shared libraries based on the full name.  This was preventing something like ts::PluginRepository::Instance()->getInput(u"file", tsp) from working when the working directory is set to the plugin directory but the executable is not found in the working directory.  This approach relies on loading shared libraries based on standard system lookup rules, but it still needs the right filename to work correctly, and the right filename in this case is "tsplugin_file", not "file".

Signed-off-by: Aaron Levinson <alevinsn_dev@levland.net>