[GSF] Syncing Fork (#7)
* Synchronous PES mode

The PES mode encapsulation supports now the Synchronous mode.

This mode includes PTS marks in the PES packets. However these marks are only useful for external tools. So, don't use this mode except if you need to remux with an external tool that **requires** to have PTS in Metadata packets. In any other case, it's a waste of space.

* Example: Reverse order of add-programinfo-id

* Use correct values

* Fixed doxygen

* Fixed spurious error message in datainject plugin

* Updated commit count to 1217 after merge

* Documentation update

* Added options --add-registration and --add-pid-registration to plugin pmt

* Added download/day in tsversion

* Added option --eit to plugin timeref

* Added latent support for packet metadata

* Added packet metadata in plugin API

* Update tsPacketEncapsulation.cpp

PacketEncapsulation: Fixed warning when building on 64-bit Windows (conversion of size_t to uint8_t)

* ApplicationSharedLibrary: Add prefix support in fallback code when loading library from filename

While the filename prefix (as in "tsplugin_") is supported in the earlier code in the constructor, it wasn't being used in the fallback code that attempts to load shared libraries based on the full name.  This was preventing something like ts::PluginRepository::Instance()->getInput(u"file", tsp) from working when the working directory is set to the plugin directory but the executable is not found in the working directory.  This approach relies on loading shared libraries based on standard system lookup rules, but it still needs the right filename to work correctly, and the right filename in this case is "tsplugin_file", not "file".

Signed-off-by: Aaron Levinson <alevinsn_dev@levland.net>

* sample-app:  Fleshed out sample application code and added an example of using a plugin pipeline using an input plugin, a single processor plugin, and an output plugin.  Also created a simple class that implements the TSP interface.

Signed-off-by: Aaron Levinson <alevinsn_dev@levland.net>

* Update sample apps with plugin API v11

* Fixed regression in shared library loading

* Added packet labels to plugins

* Documentation update

* Added option --every to plugin filter

* More accurate identification of nullified packets

* Documentation update

* Added option --reset-label on plugin filter

* Fixed race condition in message queue

* Documentation update

* Tentative support of DVB-S2 multistream on Linux

* `dos2unix` is required for building on macOS

* Implemented DVB-S2 multistream command line interface

* DVB-S2 multistream final implementation on Linux

* Added craft as packet processor plugin (untested yet)

* Fixed MSVC issue on 32-bit debug version

* Added some unit tests

* Added option --pack-pes-header in plugin craft

* Added compiling error when using GCC versions 6 and 7

* Updated doxygen

* Added option --pattern to plugni filter

* Fixed stupid typo in plugin craft