PCRAnalyzer:  Add instantaneousBitrate188() and instantaneousBitrate204() methods, misc
Comments:
-- tsPCRAnalyzer.cpp / .h:
a) Add the instantaneousBitrate188() and instantaneousBitrate204() methods, which can be used to retrieve the associated bitrates for the last second of the transport stream.  feedPacket() was adjusted to keep track of a rolling set of PCR values and packet indexes.  Made other related minor adjustments as appropriate.
b) Fixed minor grammar issue in one comment and enhanced another comment.

Signed-off-by: Aaron Levinson <alevinsn_dev@levland.net>