DektecOutputPlugin: Wait if FIFO is full in send()
tsDektecOutputPlugin.cpp:  If no longer in the starting phase, it is possible for the caller of send() to deliver packets to the plug-in faster than the Dektec output device outputs them, based on the configured bit rate.  In this case, the FIFO will fill up faster than the Dektec output device removes packets from it.  To address this potential situation and fulfill the real-time plug-in promise, which the Dektec output plug-in indicates, it now checks the FIFO load prior to calling Write() on the Dektec output channel, and if there isn't enough room, it sleeps for a short time and then tries again.  Prior to making this change, it was possible to overflow the transmit FIFO, as described in the Dektec DTAPI documentation.

Signed-off-by: Aaron Levinson <alevinsn_dev@levland.net>