DektecOutputPlugin: Add --preload-fifo-percentage and --preload-fifo-delay options for configuring the percentage of the FIFO to preload before starting transmission or the amount of media to preload into the FIFO before starting transmission based on a delay and the current bit rate, respectively.  Misc
Comments:
-- tsDektecOutputPlugin.cpp / .h:
a) Fixed some spelling errors in the help text
b) Enhanced help text documentation for --preload-fifo option to make it clear that there is no benefit to using this option if the use of the plug-in will cause the FIFO to be filled faster than real-time (or rather, faster than the bit rate that is set).
c) Added --preload-fifo-percentage option, which may be used to control the amount of FIFO to be preloaded, based on a percentage of the total FIFO size, prior to starting transmission.  Previously, this was hardcoded to always use 80%.  Now, the default is 80%.
d) Added --preload-fifo-delay option, which may be used to control the amount of FIFO to be preloaded, based on the current bit rate and delay specified, in milliseconds.  The value that the plug-in uses for the preload FIFO size, as calculated from the delay, is determined dynamically and can change if the bit rate changes. The implementation correctly handles the case that the initial bit rate is 0 and is later updated when tsp->bitrate() is called during send().
e) Enhanced verbose logging slightly.

Signed-off-by: Aaron Levinson <alevinsn_dev@levland.net>