Fixed workflow
Strange: On a local Ubuntu VM with jq-1.5-1-a5b5cbe, jq without parameter is ok,
it acts as a json pretty printer. On an Ubuntu GitHub Action runner with the same
jq-1.5-1-a5b5cbe, it fails. A dot filter (identity) must be specified. Different build
parameters on the same commit of jq?