Don't detect digital cable as ATSC
This code was from an earlier contribution of mine, but it was a mistake. 
It caused to some DVB-C tuners to indicate support for ATSC.
Within two years we weren't able to find any user having one of the rare ClearQam tuners, so I can't say how these behave and can be identified. At least, not like this.