Simplify parameters evaluation for DVB-T2 on dektec output plugin
Seems my initial algorithm was trying to be too precise, which resulted
in sometimes too long of a waiting time before streaming started.
Depending on a stream's bitrate it could go up to 10 seconds.
Let's cut off some of that precision and get the duration of the
evaluation be bitrate agnostic, so that it takes about 2.8 seconds on same
computer.