Try workaround for intermittent SSL issue during tests
On Unix systems, curl (and libcurl) have intermittent issues with
SSL/TLS connections on tsduck.io. This is probably a problem at the
web hosting facility. The problem appears from time to time, last
a few days, and disappears, until next time. This problem is not
seen with browsers and on Windows, only with curl and libcurl.

The unitary tests and the test suite download a few test samples
from tsduck.io. During these issues, tests fail and the GitHub CI/CD
workflow fails.

It has been observed that, if you make a few initial requests to
tsduck.io, some of them failing, you reduce (but not eliminate) the
probability of getting a problem during the tests.

Quite ugly, waiting for a better analysis...