EIT: change options to independantly select the 4 kinds
The current options only allow to select either actual or other, and either pf or schedule.
You have no way for instance to select actual pf, other pf and actual shedule but not other schedule.
However this is what you are supposed to do on eg. the French DTT.
This commit makes 4 distinct options, one for each kind.