Refactored implementation of IP addresses
IPAddress is now a generic class which can hold IPv4 or IPv6 addresses.
The former classes IPv4Address and IPv6Address are still there for some
time until we can remove them. They are now subclasses of IPAddress.

This commit is the base of IPv6 support. General IPv6 support is NOT
available yet but this is the start of something.

A utility tsnet has been used to debug ot test network features.