Full refactoring of PSI Repository
TSDuck was too DVB-centric. All non-DVB descriptors had to use some
horrible internal "fake" PDS. Additionally, the concept of "MPEG
private descriptor", based on registration descriptors, was not
handled. This was an obstacle to the introduction of some new
signalization. This necessary change was postponed several times,
due to its complexity (basically all signalization files and API's
had to be broken). This is now done.