tsArgMix: fix cast error on Windows
tsArgMix.cpp:95:28: error: cannot initialize return object of type 'const char *' with an rvalue of type 'const value_type *' (aka 'const wchar_t *')
[build]    95 |                     return _value.path->c_str();
[build]       |                            ^~~~~~~~~~~~~~~~~~~~
[build] 1 error generated.

The "if constexpr" will be false and not use that code (wchar_t is used).
But the code is compiled anyway and needs to compile properly.
A C-style cast solve the issue of mismatched type.