Add Unicable support to tsp dvb (input) plugin.
Add support for Unicable I and II to linux's version of tsp (EN50494 and
EN50607 respectively).

Modify some of the documentation about "satellite-number" to clarify which
hardware has what ranges. In particular, as we are adding non-DiSEqC
equipment to what is supported clarify the "non-DiSEqC" switches with
a satellite number range of 0..1 to being DiSEqC tone-burst switches.

Also clarify that what is currently referred to as a "DiSEqC switch" is
usually a "DisEqC 1.0" switch.

Test setup:

A server running Ubuntu 24.04, with a TBS dual tuner satellite card
connected via adapter 1 to a Johansson 9775 Unicable multiswitch that
in turn is connected to four satellites via wideband LNBs.

Testing:

Both Unicable I and Unicable II code paths successfully tested while
blind-scanning the entire frequency/polarity range of all four satellites
attached to the Unicable switch.