Started TLS implementation.
TLS will be used to optionally secure some TSDuck server connections
such as control ports for instance. TLS uses OpenSSL on UNIX systems
and SChannel on Windows systems. The OpenSSL implementation is done,
not the SChannel one.

Currently, TLSServer and TLSConnection are dormant classes. They are
in development but not used in any TSDuck application or plugin
(except test code in tsdebug).