Add control of DiSEqC 1.1 (uncommited) switch
Add sending a port-group 1 command, followed by a 100ms delay,
and move the tone-burst command after the port-group 0 command.

This lines up with the DisEqC bus specification sections 5.3
and 5.4.

satellite-number is not in the range 0..63. The DiSEqC 1.1 switch
moves on every four satellite-numbers, and a cascaded DisEqC 1.0
switch then moves on within those four satellite numbers.

If no suitable switch is present, then the command is ignored.

Testing:

The receiver was attached to a SPAUN SUR420 WSG DiSEqC 1.1
switch, set to mode 1;
1. port 1 had a voltage indicator attached;
2. port 2 had a DS-04C DiSEqC 1.0 switch, with voltage indicators on:
   lnb1;
   lnb2; and
   lnb3.
3. ports 3 and 4 were not connected.

Attempts were made to tune to:
* no defined satellite number (spaun port 1);
* satellite-number 0 (spaun port 1);
* satellite-number 1 (spaun port 1);
* satellite-number 4 (DS-04C lnb1);
* satellite-number 5 (DS-04C lnb2);
* satellite-number 6 (DS-04C lnb3);
* satellite-number 8 (no indicators lit).

the above were all as expected.

It was noticed that the lnb voltage is not cleared to 0V if tsp is
no longer running.