Turn of LNB power on exit
tsduck issue #1656.

Before closing the file descriptor to the satellite tuner,
call an IOCTL to turn the LNB power OFF.

Testing:

Placed a satellite finder with a power indicator in-line between the
receiver and the LNB.
Observed behaviour before change: first time tsp is run the
power comes on, and stays on.
Observed behaviour after change: The power comes on during
the time that tsp is being run, and then turns off.